modifier updeateBottle.php

+++ app/Http/Livewire/SingleCellar.php
@@ -13,6 +13,17 @@ class SingleCellar extends Component
     public $cellarId;
     public $cellar;
     public $count;
+    public $note;
+    public $qty;
+    public $bottle_qty = 0;
+    public $errorMessage;
+
+    protected $messages = [
+        'qty.required' => 'Le champ qty est obligatoire.',
+        'qty.numeric' => 'Le champ qty doit être une chiffre.',
+        'qty.min' => 'Le champ qty ne doit pas être moins :min .',
+    ];
+
     protected $listeners = ['bottleDeleted' => 'handleBottleDeleted', "incrementListen" => "increment"];
 
     public function handleBottleDeleted()
@@ -48,41 +59,52 @@ class SingleCellar extends Component
         }
     }
 
-    public function decrement($bottle_id)
+    public function updated()
     {
+        $this->validateOnly("qty", [
+            'qty' => 'required|numeric|min:0'
+        ]);
+    }
+
+
+    public function decrement($bottle_id, $bottle_quantity)
+    {
+
+        if ($this->qty > $bottle_quantity) {
+
+            session()->flash('message',  'Le champ qty ne doit pas dépasser ' . $bottle_quantity);
+            return;
+        }
+        
         $bottleInCellar = BottleInCellar::where('cellar_id', $this->cellarId)
             ->where('bottle_id', $bottle_id)
             ->first();
 
         if ($bottleInCellar) {
-            $bottleInCellar->quantity -= 1;
-          
+            $bottleInCellar->quantity -= $this->qty;
+
             if ($bottleInCellar->quantity <= 0) {
-                    $bottleInCellar->delete();
-                //     session()->flash('message', 'bouteille supprimer avec succès.');
-                // }else {
-                //     session()->flash('message', 'bouteille supprimer avec échoué.');
-                // }
-        
+                $bottleInCellar->delete();
+
+
                 $this->emit('bottleDeleted');  // emit an event to notify that a bottle was deleted
                 $bottleInCellar->quantity = 0;
             }
 
-            $bottleInCellar->save();
-            $userId = Auth::check() ? Auth::id() : null;
-
             BottleConsumed::create([
-                'bottle_id'=>  $this->bottle_id,
-                'cellar_id'=> $this->cellar_id,
+                'bottle_id' =>  $bottle_id,
+                'cellar_id' => $this->cellarId,
                 'note' => $this->note,
-                'user_id' => $userId
+                'qty' => 1,
+                "consumption_date" => date('Y-m-d'),
             ]);
+
+            $bottleInCellar->save();
             $this->reset('note');
-            
         }
     }
     // public function submitConsumed(){
-       
+
     // }
 
     // Recupère l'id dans le URL de la page directement à l'ouverture
